Digital projector feature added. Select a video and it will be displayed on the main monitor (TODO: address this) and the camera will advance in sync, as with a regular optical printer.

+++ app/lib/display/index.js
@@ -88,7 +88,7 @@ async function end () {
 	if (system.platform !== 'nix') {
 		await wv.close();
 	} else {
-		cp.kill()
+		if (cp) cp.kill()
 	}
 }
